+++ data/persistence/blob_reposiotry.py
@@ -29,6 +29,12 @@ def get_all_by_ids(session: Session, blob_ids) -> list[Blob]:
 
 
 @transactional
+def get_all_retired(session: Session) -> list[Blob]:
+    result = session.query(Blob).filter(and_(Blob.terminated.is_(None), Blob.league_id.is_(None))).all()
+    return result
+
+
+@transactional
 def get_all_by_league_order_by_id(session: Session, league_id: int) -> dict[int, Blob]:
     results = session.query(Blob).filter(Blob.league_id == league_id).all()
     return {result.id: result for result in results}